refactor(ui): Extract vueflow as a new component "LowCodeEditor" from EditorView.vue
